backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pyjwt
python-dotenv
slowapi
